--- React Project/src/component/Singup.js
@@ -1,54 +1,0 @@
-import React, { useState } from 'react';
-import './SignupForm.css';
-
-function SignupForm() {
-  const [email, setEmail] = useState('');
-  const [password, setPassword] = useState('');
-  const [confirmPassword, setConfirmPassword] = useState('');
-
-  const handleSubmit = (e) => {
-    e.preventDefault();
-    console.log('Submitted:', { email, password, confirmPassword });
-  };
-
-  return (
-    <div className="container">
-      <h2>회원가입</h2>
-      <form onSubmit={handleSubmit}>
-        <label htmlFor="email">이메일:</label>
-        <input
-          type="email"
-          id="email"
-          name="email"
-          value={email}
-          onChange={(e) => setEmail(e.target.value)}
-          required
-        />
-
-        <label htmlFor="password">비밀번호:</label>
-        <input
-          type="password"
-          id="password"
-          name="password"
-          value={password}
-          onChange={(e) => setPassword(e.target.value)}
-          required
-        />
-
-        <label htmlFor="confirm_password">비밀번호 확인:</label>
-        <input
-          type="password"
-          id="confirm_password"
-          name="confirm_password"
-          value={confirmPassword}
-          onChange={(e) => setConfirmPassword(e.target.value)}
-          required
-        />
-
-        <input type="submit" value="가입하기" />
-      </form>
-    </div>
-  );
-}
-
-export default SignupForm;
